Moved NotesDirPath class to utils
